purple-agents/advanced/requirements.txt
google-adk[a2a]>=0.1.0
google-genai>=1.0.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
a2a>=0.1.0
